change the test file to use the providers code
